% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellcyclegenes_whitfield2002}
\alias{cellcyclegenes_whitfield2002}
\title{List of Cell Cycle Genes Identified in Whitfield et al 2002.}
\format{
A list with the following elements:

\describe{
  \item{hgnc}{Gene symbol}
  \item{ensembl}{ENSEMBL gene ID}
  \item{phase}{Marker phase identified in Whitfield et al. 2002}
 }
}
\description{
List of cell cycle genes and their associated cell
  cycle state as reported in Whitfield et al. 2002.
}
\examples{
data(cellcyclegenes_whitfield2002)

}
\keyword{data}
