% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_dist.R
\name{circ_dist}
\alias{circ_dist}
\title{Pairwise distance between two circular variables}
\usage{
circ_dist(y1, y2)
}
\arguments{
\item{y1}{A vector of angles.}

\item{y2}{A vector of angles.}
}
\value{
A vector of distances between angles.
}
\description{
We define distance between two angles: the minimum of
the differences in both clockwise and counterclockwise directions.
}
\author{
Joyce Hsiao, Matthew Stephens
}
