% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mask.R
\name{create_mask}
\alias{create_mask}
\title{Fluorescence image analysis}
\usage{
create_mask(bright, dapi, gfp, rfp, id, size_cutoff = 350,
  display = TRUE, write = TRUE)
}
\description{
generate a mask surrounding nuclei
takes all data channels, uses dapi to create a mask,
then returns a smaller image containing each nucleus
and writes them to files
it returns a data frame with information about the cells
I reccomend changing the file output directories, etc to make it work
better in the workflow.
}
\author{
Kenneth Barr
}
