% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.estep}
\alias{cycle.npreg.estep}
\title{Compute estimates in one iteration}
\usage{
\method{cycle}{npreg.estep}(Y, theta, mu_est, sigma_est, nbins = 100, ncores,
  ...)
}
\arguments{
\item{Y}{gene by sample log2CPM gene expression matrix}

\item{theta}{cell time estimate to be used in estimating cyclical patterns
of gene exprssion levels for each gene}

\item{ncores}{number of cores to be used}
}
\description{
Given the estimated cell times, gene mean and gene variation,
  (possibly from m-step), compute the expected cell times at midpoint of defined bins
}
