% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.insample}
\alias{cycle.npreg.insample}
\title{Estimate cell cycle ordering in the current sample}
\usage{
\method{cycle}{npreg.insample}(Y, theta, ncores = 12, polyorder = 3,
  method.trend = c("npcirc.nw", "npcirc.ll", "trendfilter", "loess",
  "bspline"), ...)
}
\arguments{
\item{update}{T/F to update cell times}
}
