% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.loglik}
\alias{cycle.npreg.loglik}
\title{log-likelihood of nonparametric smoothing}
\usage{
\method{cycle}{npreg.loglik}(Y, theta, mu_est, sigma_est)
}
\arguments{
\item{Y}{gene by sample expression matrix}

\item{mu_est}{gene by sample matrix of expected mean}

\item{sigma_est}{vector of standard errors for each gene}
}
