% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.outsample.noiter}
\alias{cycle.npreg.outsample.noiter}
\title{Predict test-sample ordering using training lables (no update)}
\usage{
\method{cycle}{npreg.outsample.noiter}(Y_test, sigma_est, funs_est,
  method.trend = c("npcirc.nw", "npcirc.ll", "trendfilter"),
  method.initialize.theta = c("pca", "uniform"), ncores = 12,
  maxiter = 10, tol = 1, verbose = TRUE, ...)
}
