% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.outsample}
\alias{cycle.npreg.outsample}
\title{Predict cell cycle ordering in the test samples}
\usage{
\method{cycle}{npreg.outsample}(Y_test, theta_est, mu_est, sigma_est,
  ncores = 12, ...)
}
