% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{cycle.npreg.unsupervised}
\alias{cycle.npreg.unsupervised}
\title{unsupervised method for learning cell times}
\usage{
\method{cycle}{npreg.unsupervised}(Y, theta, ncores = 12,
  method.trend = c("npcirc.nw", "npcirc.ll", "trendfilter", "loess",
  "bspline"), grids = 100, polyorder = 2, maxiter = 30, verbose = TRUE,
  tol = 1, ...)
}
