% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.spml.R
\name{cycle.spml.testmodel}
\alias{cycle.spml.testmodel}
\title{Estimate gene weights for cell time}
\usage{
\method{cycle}{spml.testmodel}(Y_test, Y_train, theta_test, theta_train)
}
\arguments{
\item{Y_test}{gene by testing samples}

\item{theta_test}{gene by training samples}

\item{theta_train}{cell times for training samples}

\item{theta_test}{cell times for test samples}
}
