% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.spml.R
\name{cycle.spml.trainmodel}
\alias{cycle.spml.trainmodel}
\title{Estimate gene weights for cell time}
\usage{
\method{cycle}{spml.trainmodel}(Y, theta)
}
\arguments{
\item{Y}{gene by sample expression matrix}

\item{theta}{sample cell time vector}
}
