% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_npreg.R
\name{cycle_npreg_insample}
\alias{cycle_npreg_insample}
\title{Obtain Cyclic Trend Estimates from Training Data}
\usage{
cycle_npreg_insample(
  Y,
  theta,
  ncores = 2,
  polyorder = 2,
  method.trend = c("trendfilter", "loess", "bspline")
)
}
\arguments{
\item{Y}{A matrix (gene by sample) of normalized and transformed
gene expression values.}

\item{theta}{A vector of angles.}

\item{ncores}{We use the doParallel package for parallel computing.}

\item{polyorder}{We estimate cyclic trends of gene expression
levels using nonparamtric trend filtering.}

\item{method.trend}{How to estimate cyclic trend of gene expression
values. We offer three options: \code{method.trend =
"trendfilter"}, uses \code{\link{fit_trendfilter}},
\code{method.trend = "loess"} uses \code{\link{fit_loess}}), and
\code{method.trend = "bsplines"} uses \code{\link{fit_bspline}}.  We
found that \code{"trendfilter"} provided the best fits in our
experiments. However, trend-filtering may require more
computational effort since it uses cross-validation, so for fast
results we recommend using \code{"bspline"}.}
}
\value{
A list with four elements:

\item{Y}{The gene expression marix.}

\item{theta}{Vector of angles or cell cycle phases.}

\item{sigma_est}{Estimated standard error of the cyclic trend for
  each gene.}

\item{funs_est}{A list of functions for approximating the cyclic
trends of gene express levels for each gene.}
}
\description{
Estimate cyclic trends of gene expression levels
  using training data.
}
\examples{
# See \code{help(cycle_npreg_outsample)} for an example.

}
\seealso{
\code{\link{cycle_npreg_mstep}} for estimating cyclic functions
  given inferred phases from \code{\link{cycle_npreg_loglik}}, and
  \code{\link{cycle_npreg_outsample}} for predicting cell cycle phase
  using parameters learned from \code{\link{cycle_npreg_insample}}.
}
\author{
Joyce Hsiao
}
