% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_npreg.R
\name{cycle_npreg_loglik}
\alias{cycle_npreg_loglik}
\title{Infer Angles or Cell Cycle Phase Using Gene Expression Data}
\usage{
cycle_npreg_loglik(
  Y,
  sigma_est,
  funs_est,
  grids = 100,
  method.grid = c("pca", "uniform")
)
}
\arguments{
\item{Y}{Gene by sample expression matrix.}

\item{sigma_est}{A vector of standard errors for each gene from the
training data.}

\item{funs_est}{A vector of cyclic functions estimated for each
gene from the training data.}

\item{grids}{Number of bins used over interval 0 to \code{2*pi}.}

\item{method.grid}{The approach used to initialize angles in the
computation. \code{method.grid = "uniform"} creates k
equally-spaced bins ("grids"). \code{method.grid = "pca"} uses gene
expression values to infer angles, and then these angles are used
to project the cells to the closest bin.}
}
\value{
A list with the following three elements:

\item{cell_times_est}{Inferred angles or cell cycle phases (not
  ordered).}

\item{loglik_est}{Log-likelihood estimates for each gene.}

\item{prob_per_cell_by_celltimes}{Probabilities of each cell belonging
  to each bin.}
}
\description{
Infer Angles or Cell Cycle Phase Using Gene Expression Data
}
\keyword{internal}
