% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eset_final}
\alias{eset_final}
\title{Filtered gene expression data, along with summarized FUCCI intensities
and sample information}
\format{An ExpressionSet object with 888 samples and 11,093 genes.
\describe{
  \item{\code{theta}}{inferred angles}
  \item{\code{exprs(obj)}}{Molecule counts}
  ...
}}
\usage{
eset_final
}
\description{
Contains
}
\keyword{datasets}
