% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eset_final_sub}
\alias{eset_final_sub}
\title{Molecule counts of top 100 cyclical genes in 888 samples analyzed
  in the study.}
\format{An ExpressionSet object with 888 samples and top 100 cyclic genes.

\describe{
  \item{\code(pData(eset_final_sub)$theta}{Inferred angles of each cell along
    a circle, also known as FUCCI phase - cell cycle phase inferred from
    FUCCI intensities.}

  \item{\code{exprs(est_final_sub)}}{Molecule counts of top 100 cyclical genes.}
}}
\description{
An ExpressionSet object (require Biobase package) including
  molecule count data after gene and smaple filtering. The phenotypeData() slot
  contains sample phenotype information and the featureData() slot contains
  gene feature information.
}
\keyword{data}
