% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cyclical.R, R/fit.trendfilter.generic.R
\name{fit.bspline}
\alias{fit.bspline}
\alias{fit.bspline}
\title{fitting bsplines}
\usage{
fit.bspline(yy, time, polyorder = 2)

fit.bspline(yy, time, polyorder = 2)
}
\arguments{
\item{yy}{log2cpm gene expression vector}

\item{yy}{log2cpm gene expression vector}
}
