% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cyclic.many.R
\name{fit.cyclical.many.pve}
\alias{fit.cyclical.many.pve}
\title{Compute proportation of variance explained by cyclic trends in
the gene expression levels for each gene.}
\usage{
fit.cyclical.many.pve(Y, theta, polyorder = 2, ncores = 4)
}
\arguments{
\item{Y}{A matrix (gene by sample) of gene expression values.
The expression values are assumed to have been normalized and
transformed to standard normal distribution.}

\item{theta}{A vector of cell cycle phase (angles) for single-cell samples.}

\item{polyorder}{We estimate cyclic trends of gene expression levels using
nonparamtric trend filtering. The default fits second degree polynomials
(polyorder=2).}

\item{ncores}{mclapply from the parallel package is used to perform
parallel computing to reduce computational time. Default is 4.}
}
\value{
A vector of proportion of variance explained for each gene.
}
\description{
We applied quadratic (second order) trend filtering using the
trendfilter function in the genlasso package (Tibshirani, 2014).
The trendfilter function implements a nonparametric smoothing method
which chooses the smoothing parameter by cross-validation and fits
a piecewise polynomial regression. In more specifics: The trendfilter
method determines the folds in cross-validation in a nonrandom manner.
Every k-th data point in the ordered sample is placed in the k-th fold,
so the folds contain ordered subsamples. We applied five-fold
cross-validation and chose the smoothing penalty using
the option lambda.1se: among all possible values of the penalty term,
the largest value such that the cross-validation standard error is within
one standard error of the minimum. Furthermore, we desired that the estimated
expression trend be cyclical. To encourage this, we concatenated the ordered
gene expression data three times, with one added after another. The quadratic
trend filtering was applied to the concatenated data series of each gene.
}
\author{
Joyce Hsiao
}
