% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cyclical.R, R/fit.trendfilter.generic.R
\name{fit.loess}
\alias{fit.loess}
\alias{fit.loess}
\title{fitting loess}
\usage{
fit.loess(yy, time, polyorder = 2)

fit.loess(yy, time, polyorder = 2)
}
\arguments{
\item{yy}{log2cpm gene expression vector}

\item{yy}{log2cpm gene expression vector}
}
