% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit}
\alias{fit}
\title{Training and prediction results}
\format{\describe{
  \item{\code{fit_train}}{training results}

  \item{\code{fit_predict}}{prediction results}
 }}
\description{
Pre-computed results, including training and
prediction and sample information.
}
\keyword{data}
