% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cyclic.one.R
\name{fit.trendfilter.generic}
\alias{fit.trendfilter.generic}
\title{Using trendfiltering to estimate cyclic trend of gene expression}
\usage{
fit.trendfilter.generic(yy, polyorder = 2)
}
\arguments{
\item{yy}{A vector of gene expression values for one gene. The expression
values are assumed to have been normalized and transformed to
standard normal distribution.}

\item{polyorder}{We estimate cyclic trends of gene expression levels using
nonparamtric trend filtering. The default fits second degree polynomials
(polyorder=2).}
}
\value{
\describe{
     \item{\code{trend.yy}}{The estimated cyclic trend.}
     \item{\code{pve}}{Proportion of variance explained by the cyclic trend
         in the gene expression levels}
         }
}
\description{
We applied quadratic (second order) trend filtering using the
trendfilter function in the genlasso package (Tibshirani, 2014).
The trendfilter function implements a nonparametric smoothing method
which chooses the smoothing parameter by cross-validation and fits
a piecewise polynomial regression. In more specifics: The trendfilter
method determines the folds in cross-validation in a nonrandom manner.
Every k-th data point in the ordered sample is placed in the k-th fold,
so the folds contain ordered subsamples. We applied five-fold
cross-validation and chose the smoothing penalty using
the option lambda.1se: among all possible values of the penalty term,
the largest value such that the cross-validation standard error is within
one standard error of the minimum. Furthermore, we desired that the estimated
expression trend be cyclical. To encourage this, we concatenated the ordered
gene expression data three times, with one added after another. The quadratic
trend filtering was applied to the concatenated data series of each gene.
}
\author{
Joyce Hsiao
}
