% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclic_one.R
\name{fit_bspline}
\alias{fit_bspline}
\title{Use bsplies to cyclic trend of gene expression levels}
\usage{
fit_bspline(yy, time)
}
\arguments{
\item{yy}{A vector of gene expression values for one gene. The expression
values are assumed to have been normalized and transformed to
standard normal distribution.}

\item{time}{A vector of angels (cell cycle phase).}
}
\value{
\describe{
     \item{\code{pred.yy}}{The estimated cyclic trend.}
         }
}
\description{
Use bsplies to cyclic trend of gene expression levels
}
\author{
Joyce Hsiao
}
