% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_pred}
\alias{fit_pred}
\title{Results of predicting cell cycle phase for samples from NA19098.}
\format{\describe{
  \item{\code{Y_reordered}}{Normalized expression values (log2CPM , )}
  \item{\code{cell_times_reordered}}{Estimated phase}
 }}
\description{
Pre-computed results. Applied \code{cycle_npreg_outsample} and
  results stored in \code{fit_train} to predict cell cycle phase for
  single-cell samples of NA19098.
}
\keyword{data}
