% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_train}
\alias{fit_train}
\title{Traing model results among samples from 5 individuals.}
\description{
Pre-computed results. Applied \code{cycle_npreg_insample} to
  obtain gene-specific cyclic trend parameters using samples from 5
  individuals
}
\keyword{data}
