% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_cyclic_list}
\alias{genes_cyclic_list}
\title{Genes ordered by their cyclic trend}
\format{A data frame with the following columns, in which the rows
are ordered according to the proportion of variance explained
(large to small):

\describe{

  \item{ensg}{ENSG gene ID.}

  \item{pve}{Proportion of variance explained in the expression
  values by the estimated cyclic trend.}
}}
\description{
Genes ordered by their cyclic trend (strong to weak),
and sample information.
}
\keyword{data}
