% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_npreg.R
\name{initialize_grids}
\alias{initialize_grids}
\title{Initialize Grid for cycle_npreg_loglik}
\usage{
initialize_grids(Y, grids = 100, method.grid = c("pca", "uniform"))
}
\arguments{
\item{Y}{Gene expression matrix (gene by sample).}

\item{grids}{Number of bins used over interval 0 to \code{2*pi}.}

\item{method.grid}{The approach used to initialize angles in the
computation. \code{method.grid = "uniform"} creates k
equally-spaced bins ("grids"). \code{method.grid = "pca"} uses gene
expression values to infer angles, and then these angles are used
to project the cells to the closest bin.}
}
\value{
A vector of initialized angles to be used in
  \code{\link{cycle_npreg_loglik}}.
}
\description{
For prediction, initialize grid points for cell cycle phase
  on a circle.
}
\keyword{internal}
