% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_5genes_predict}
\alias{model_5genes_predict}
\title{Cell-Cycle Predictions using 5 Genes}
\format{
A \code{SingleCellExperiment} object. The predicted cell
cycle is stored in \emph{model_5genes_predict$cellcycle_peco}.
}
\description{
Pre-computed peco results obtained by applying
  \emph{cycle_npreg_outsample} with model stored in
  \emph{model_5genes_train} to predict cell cycle phase for
  single-cell samples of NA19098.
}
\examples{
data(model_5genes_predict)

}
\keyword{data}
