% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_5genes_train}
\alias{model_5genes_train}
\title{Model Training Results for 5 Genes}
\format{
A list with the follwing elements:
\describe{

\item{Y}{A matrix (gene by sample) of quantile-normailzed gene
  expression values.}

\item{theta}{A vector of cell cycle phase values (ranging between
  0 and \code{2*pi}).}

\item{sigma_est}{A vector of estimated standard errors, one for
  each gene.}

\item{funs_est}{A list of estimated cyclic functions.}}
}
\description{
Pre-computed gene-specific cyclic trend results
  obtained by applying \code{cycle_npreg_insample}.
}
\examples{
data(model_5genes_train)

}
\keyword{data}
