% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{msd}
\alias{msd}
\title{Mean squared deviation from the predicted cyclical trend}
\usage{
msd(Y, mu_est)
}
\arguments{
\item{Y}{gene by sample expression matrix}

\item{mu_est}{Fitted cyclical trend per gene for each sample}
}
