% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionSamples.R
\name{partitionSamples}
\alias{partitionSamples}
\title{partition samples by specified sizes of training and testing data}
\usage{
partitionSamples(y, runs, nsize.each = NULL)
}
\arguments{
\item{y}{data vector}

\item{runs}{number of times that data vector will be partitioned}

\item{nsize_each}{number of samples in each partition}
}
