% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{primes}
\alias{primes}
\title{Print prime numbers}
\usage{
primes(nprimes, prime.start = 1)
}
\arguments{
\item{nprimes}{number of prime numbers needed (default starting from 1)}

\item{prime.start}{start search for prime numbers from this value}
}
