% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{pve}
\alias{pve}
\title{Proportion of variance explained by the fitted cyclical trend}
\usage{
pve(Y, mu_est)
}
\arguments{
\item{Y}{gene by sample expression matrix}

\item{mu_est}{Fitted cyclical trend per gene for each sample}
}
