% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rFL.CorrCoeff}
\alias{rFL.CorrCoeff}
\title{Fisher-Lee correlation coefficient}
\usage{
rFL.CorrCoeff(lcdat1, lcdat2)
}
\arguments{
\item{lcdat1}{length n vector of radians}

\item{lcdat2}{length n vector of radians}
}
\references{
Fisher NJ and Lee AJ (1983). A correlation coefficient for circular data. Biometrika, 70, 327-332.
   Pewsey et al. Circular statistics in R
}
