% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rFL.IndTestRand}
\alias{rFL.IndTestRand}
\title{statistical significance of Fisher-Lee correlation coefficient, permutation-based}
\usage{
rFL.IndTestRand(lcdat1, lcdat2, NR = 10000)
}
\arguments{
\item{lcdat1}{length n vector of radians}

\item{lcdat2}{length n vector of radians}

\item{NR}{number of permutations}
}
\references{
Fisher NJ and Lee AJ (1983). A correlation coefficient for circular data. Biometrika, 70, 327-332.
   Pewsey et al. Circular statistics in R
}
