% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rFLCorrCoeff}
\alias{rFLCorrCoeff}
\title{compute Fisher-Lee correlation coefficient}
\usage{
rFLCorrCoeff(lcdat1, lcdat2)
}
\arguments{
\item{lcdat1}{length n vector of radians}

\item{lcdat2}{length n vector of radians}
}
\references{
Pewsey et al. Circular statistics in R
}
