% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rFLIndTestRand}
\alias{rFLIndTestRand}
\title{compute statistical significance of Fisher-Lee correlation coefficient}
\usage{
rFLIndTestRand(lcdat1, lcdat2, NR)
}
\arguments{
\item{lcdat1}{length n vector of radians}

\item{lcdat2}{length n vector of radians}

\item{NR}{number of permutations}
}
\references{
Pewsey et al. Circular statistics in R
}
