% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rFLRank.CorrCoeff}
\alias{rFLRank.CorrCoeff}
\title{Circular-circular rank correlation coefficient}
\usage{
rFLRank.CorrCoeff(lcdat1, lcdat2)
}
\references{
Fisher NJ and Lee AJ (1983). A correlation coefficient for circular data. Biometrika, 70, 327-332.
}
