% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rFLRank.IndTestRand}
\alias{rFLRank.IndTestRand}
\title{Significance test for circular rank correlation, permutation-based}
\usage{
rFLRank.IndTestRand(lcdat1, lcdat2, NR = 10000)
}
\arguments{
\item{lcdat1}{length n vector of radians}

\item{lcdat2}{length n vector of radians}

\item{NR}{number of permutations}
}
\references{
Pewsey et al. Circular statistics in R
}
