% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoFLCIBoot.R
\name{rhoFLCIBoot}
\alias{rhoFLCIBoot}
\title{compute boostrap confidence interval for correlation coefficient}
\usage{
rhoFLCIBoot(lcdat1, lcdat2, ConfLevel, B)
}
\arguments{
\item{lcdat1}{length n vector of radians}

\item{lcdat2}{length n vector of radians}

\item{ConfLevel}{percent confidence interval}

\item{B}{number of bootstrap sample}
}
\references{
Pewsey et al. Circular statistics in R
}
