% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.corr.R
\name{rotation}
\alias{rotation}
\title{Compute correlations between rank orders of the cell times}
\usage{
rotation(y1, y2, ...)
}
\arguments{
\item{y1}{numeric vector of length N with values betwen 0 to 2pi}

\item{y2}{numeric vector of length N with values betwen 0 to 2pi}
}
