% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seurat.R
\name{run_seurat}
\alias{run_seurat}
\title{Seurat code for cell cycle scoring}
\usage{
run_seurat(Y, s.genes, g2m.genes, n.bin = 25, seed.use = 1,
  random.seed = 1)
}
\arguments{
\item{Y}{log2 normalized gene expression matrix}
}
\description{
Compute phase specific cell cycle scores as outlined in Seurat
}
