% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_origin.R
\name{shift_origin}
\alias{shift_origin}
\title{Shift origin of the angles}
\usage{
shift_origin(phase, origin)
}
\arguments{
\item{phase}{A vector of angles (in radians).}

\item{origin}{the new origin of the angles.}
}
\value{
A vector of angles shifted to the new origin.
}
\description{
Shift origin of the angles for visualization
}
\author{
Joyce Hsiao
}
