% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_human}
\alias{training_human}
\title{Training Data from 888 Single-cell Samples and 101 Top
  Cyclic Genes}
\format{
A list with the follwing elements:

\describe{
  \item{predict.yy}{Estimated cyclic expression values in the
training data}

  \item{cellcycle_peco_ordered}{Training labels ordered from 0 to
    2*pi}

  \item{cell_cycle function}{Nonparametric function of cyclic gene
expression trend obtained by trendfilter function in genlasso}

  \item{pve}{Proportion of variance explained in each gene by the
  cell cycl phase label}
 }
}
\usage{
data(training_human)
}
\description{
Pre-computed results. Applied \code{fit_cyclic_many}
to 888 single-cell samples that have both normalized gene
expression values and cell cycle labels to obtain training results
that can be used as input for predicting cell cycle phase in other
data.
}
\keyword{data}
