% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellcyclegenes_whitfield2002}
\alias{cellcyclegenes_whitfield2002}
\title{List of Cell Cycle Genes Identified in Whitfield et al, 2002.}
\format{
A data frame with the three columns:

\describe{
  \item{hgnc}{HGNC gene symbol.}
  \item{ensembl}{Ensembl gene identifier.}
  \item{phase}{Marker phase identified in Whitfield \emph{et al} (2002).}
 }
}
\description{
List of cell cycle genes and their associated cell
  cycle state from Whitfield \emph{et al} (2002).
}
\examples{
data(cellcyclegenes_whitfield2002)

}
\keyword{data}
