% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_dist.R
\name{circ_dist}
\alias{circ_dist}
\title{Pairwise Distance Between Two Circular Variables}
\usage{
circ_dist(y1, y2)
}
\arguments{
\item{y1}{A vector of angles.}

\item{y2}{A vector of angles.}
}
\value{
A vector of distances between the angles.
}
\description{
We define distance between two angles as the minimum of
  the differences in both clockwise and counterclockwise directions.
}
\examples{
# Vector of angles.
theta_ref <- seq(0,2*pi, length.out=100)

# Shift the origin of theta_ref to pi.
theta_compare <- shift_origin(theta_ref, origin = pi)
mean(circ_dist(theta_ref, theta_compare))

# After rotation of angles, difference is zero between the original
# and the shifted angles.
theta_compare_rotated <- rotation(ref_var=theta_ref,
    shift_var=theta_compare)
mean(circ_dist(theta_ref, theta_compare_rotated))

}
\author{
Joyce Hsiao, Matthew Stephens
}
