% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.loglik}
\alias{cycle.npreg.loglik}
\title{log-likelihood of nonparametric smoothing}
\usage{
\method{cycle}{npreg.loglik}(Y, sigma_est, funs_est, grids = 100,
  method.type = c("supervised", "unsupervised"), method.grid = c("pca",
  "uniform"), ...)
}
\arguments{
\item{Y}{gene by sample expression matrix}

\item{sigma_est}{vector of standard errors for each gene}

\item{mu_est}{gene by sample matrix of expected mean}

\item{Y}{gene by sample expression matrix}

\item{mu_est}{gene by sample matrix of expected mean}

\item{sigma_est}{vector of standard errors for each gene}
}
