% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{cycle.npreg.mstep}
\alias{cycle.npreg.mstep}
\title{Estimate parameters for the cyclial ordering using nonparametric smoothing}
\usage{
\method{cycle}{npreg.mstep}(Y, theta, ncores = 12, ...)
}
\arguments{
\item{Y}{gene by sample expression matrix (log2CPM).}

\item{theta}{observed cellt times}
}
\description{
Conditioned on observed cell times, estimate the cyclical trend
  for each gene, and compute the likelihood of the observed cell times.
}
