% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.npreg.R
\name{cycle.npreg.mstep}
\alias{cycle.npreg.mstep}
\title{Estimate parameters of the cyclic trends}
\usage{
\method{cycle}{npreg.mstep}(Y, theta, method.trend = c("trendfilter",
  "npcirc.nw", "npcirc.ll", "loess", "bspline"), polyorder = 2,
  ncores = 4, ...)
}
\arguments{
\item{Y}{gene by sample expression matrix (log2CPM).}

\item{theta}{observed cellt times}
}
\value{
\describe{
    \item{\code{Y}{Input gene expression data}}
    \item{\code{theta}{Input angles}}
    \item{\code{mu_est}{Estimated expression levels given the cyclic function for each gene}}
    \item{\code{sigma_est}{Estimated standard error of the cyclic trends for each gene}}
    \item{\code{funs_est}{Estimated cyclic functions}}
    }
}
\description{
This is used in both cycle.npreg.insample (training data fitting)
and cycle.npreg.outsample (testing data prediction) to estimate cyclic
trends of gene expression values. The function outputs for each gene
standard error of the cyclic trend, cyclic function, and the estimated
expression levels given the cyclic function.
}
\author{
Joyce Hsiao
}
