% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_npreg.R
\name{cycle_npreg_insample}
\alias{cycle_npreg_insample}
\title{Obtain cyclic trend estimates from the training data}
\usage{
cycle_npreg_insample(Y, theta, ncores = 4, polyorder = 2,
  method.trend = c("trendfilter", "loess", "bspline"))
}
\arguments{
\item{Y}{A matrix of normalized and transformed gene expression
values.  Gene by sample.}

\item{theta}{A vector of angles.}

\item{ncores}{We use doParallel package for parallel computing.}

\item{polyorder}{We estimate cyclic trends of gene expression
levels using nonparamtric trend filtering. The default fits second
degree polynomials.}

\item{method.trend}{Varous methods that can be applied to estimate
cyclic trend of gene expression levels.}
}
\value{
A list with four elements:

\item{Y}{Gene expression marix.}

\item{theta}{Vector of angles or cell cycle phases.}

\item{sigma_est}{Estimated standard error of the cyclic trend for
each gene.}

\item{funs_est}{A list of functions for approximating the cyclic
trends of gene express levels for each gene.}
}
\description{
Estimates cyclic trends of gene expression levels
using training data.
}
\examples{
\dontrun{
# import data
data(sce_top101genes)

# select top 5 cyclic genes
sce_top5 <- sce_top101genes[
    order(rowData(sce_top101genes)$pve_fucci, decreasing=TRUE)[1:5],]

# reordering the data according to FUCCI phase
coldata <- colData(sce_top5)[order(colData(sce_top5)$theta_shifted),]

# quantile-transform CPM to normal distribution
sce_quant <- data_transform_quantile(sce_top5)
expr_quant <- assay(sce_quant, "cpm_quant")

# Select samples from NA18511 for our prediction example
which_samples_train <- rownames(coldata)[coldata$chip_id != "NA18511"]
which_samples_predict <- rownames(coldata)[coldata$chip_id == "NA18511"]

# make an example of using data from 5 individuals to predict phase
# in one indivdual
Y_train <- expr_quant[, colnames(expr_quant) \%in\% which_samples_train]
theta_train <-
    coldata$theta_shifted[rownames(coldata) \%in\% which_samples_train]
names(theta_train) <-
    rownames(coldata)[rownames(coldata) \%in\% which_samples_train]

# obtain cyclic function estimates
model_5genes_train <- cycle_npreg_insample(Y = Y_train,
                                  theta = theta_train,
                                  polyorder=2,
                                  ncores=2,
                                  method.trend="trendfilter")

Y_predict <- expr_quant[, which_samples_predict]

theta_test <- coldata$theta[which_samples_predict]
names(theta_test) <- rownames(coldata)[which_samples_predict]

model_5genes_predict <- cycle_npreg_outsample(Y_test=Y_predict,
                                    sigma_est=model_5genes_train$sigma_est,
                                    funs_est=model_5genes_train$funs_est,
                                    method.trend="trendfilter",
                                    ncores=1,
                                    get_trend_estimates=TRUE)

par(mfrow=c(2,3), mar=c(4,4,3,1))
for (g in seq_len(5)) {
  plot(model_5genes_predict$Y_reordered[g,],
       x=model_5genes_predict$cell_times_reordered, axes=FALSE,
       xlab="FUCCI phase",
       ylab="Predicted phase")
  points(y=model_5genes_predict$funs_reordered[[g]](
      model_5genes_predict$cell_times_reordered),
         x=model_5genes_predict$cell_times_reordered,
         pch=16, col="royalblue")
  axis(2);
  axis(1,at=c(0,pi/2, pi, 3*pi/2, 2*pi),
           labels=c(0,expression(pi/2), expression(pi), expression(3*pi/2),
                    expression(2*pi)))
  abline(h=0, lty=1, col="black", lwd=.7)
  title(rownames(model_5genes_predict$Y_reordered)[g])
}
title("Predicting cell cycle phase for NA18511", outer=TRUE)
}

}
\seealso{
\code{\link{cycle_npreg_mstep}} for estimating cyclic functions given
    inferred phases from  \code{\link{cycle_npreg_loglik}},
    \code{\link{cycle_npreg_outsample}} for predicting cell cycle phase
     using parameters learned from \code{\link{cycle_npreg_insample}}

Other peco classifier functions: \code{\link{cycle_npreg_loglik}},
  \code{\link{cycle_npreg_mstep}},
  \code{\link{cycle_npreg_outsample}},
  \code{\link{initialize_grids}}
}
\author{
Joyce Hsiao
}
\concept{peco classifier functions}
