% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_npreg.R
\name{cycle_npreg_insample}
\alias{cycle_npreg_insample}
\title{Obtain cyclic trend estimates from the training data}
\usage{
cycle_npreg_insample(Y, theta, ncores = 4, polyorder = 2,
  method.trend = c("trendfilter", "loess", "bspline"), ...)
}
\arguments{
\item{Y}{A matrix of normalized and transformed gene expression values.
Gene by sample.}

\item{theta}{A vector of angele}

\item{ncores}{We use mclapply function for parallel computing.
Default is 4 cores.}

\item{polyorder}{We estimate cyclic trends of gene expression levels using
nonparamtric trend filtering. The default fits second degree polynomials
(polyorder=2).}

\item{method.trend}{Varous methods that can be applied to estimate cyclic trend
of gene expression levels.}
}
\value{
\describe{
     \item{\code{Y}}{Gene expression marix.}
     \item{\code{theta}}{Vector of angles or cell cycle phases}
     \item{\code{sigma_est}}{Estimated standard error of the cyclic trend for each gene.}
     \item{\code{funs_est}}{A list of functions for approximating the cyclic trends of gene express levels for each gene.}
         }
}
\description{
Estimates cyclic trends of gene expression levels using training data
}
\author{
Joyce Hsiao
}
