% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_npreg.R
\name{cycle_npreg_mstep}
\alias{cycle_npreg_mstep}
\title{Estimate parameters of the cyclic trends}
\usage{
cycle_npreg_mstep(Y, theta, method.trend = c("trendfilter", "loess",
  "bspline"), polyorder = 2, ncores = 4, ...)
}
\arguments{
\item{Y}{Gene by sample expression matrix (log2CPM).}

\item{theta}{Observed cell times.}
}
\value{
A list with the following elements: 
    
\item{Y}{Input gene expression data.}

\item{theta}{Input angles.}

\tem{mu_est}{Estimated expression levels given the cyclic function
for each gene.}

\item{sigma_est}{Estimated standard error of the cyclic trends for
each gene.}

\item{funs_est}{Estimated cyclic functions.}
}
\description{
This is used in both cycle_npreg_insample (training
data fitting) and cycle_npreg_outsample (testing data prediction)
to estimate cyclic trends of gene expression values. The function
outputs for each gene standard error of the cyclic trend, cyclic
function, and the estimated expression levels given the cyclic
function.
}
\author{
Joyce Hsiao
}
