% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_npreg.R
\name{cycle_npreg_outsample}
\alias{cycle_npreg_outsample}
\title{Predict test-sample ordering using training labels (no update)}
\usage{
cycle_npreg_outsample(Y_test, sigma_est, funs_est,
  method.trend = c("trendfilter", "loess", "bspline"), polyorder = 2,
  method.grid = "uniform", ncores = 4, grids = 100,
  get_trend_estimates = F)
}
\arguments{
\item{Y_test}{Gene expression data to be used for prediction.
Gene by sample.}

\item{sigma_est}{A vector of gene-specific standard error of the
cyclic trends.}

\item{funs_est}{A vector of cyclic functions estimating cyclic trends.}

\item{method.trend}{Varous methods that can be applied to estimate
cyclic trend of gene expression levels.}

\item{polyorder}{We estimate cyclic trends of gene expression levels using
nonparamtric trend filtering. The default fits second degree polynomials.}

\item{method.grid}{Method for defining bins along the circle.}

\item{ncores}{We use mclapply function for parallel computing.}

\item{grids}{number of bins to be selected along 0 to 2pi.}

\item{get_trend_estimates}{To re-estimate the cylic trend based on the predicted
cell cycle phase or not (T or F). Default FALSE. This step calls trendfilter
and is computationally intensive.}
}
\value{
A list with the following elements:

\item{Y}{The input gene expression marix.}

\item{cell_times_est}{Inferred angles or cell cycle phases, NOT
ordered.}

\item{loglik_est}{Log-likelihood estimates for each gene.}

\item{cell_times_reordered}{The inferred angles reordered (in
ascending order).}

\item{Y_reorded}{The input gene expression matrix reordered by
cell_times_reordered.}

\item{sigma_reordered}{Estimated standard error of the cyclic trend
for each gene, reordered by cell_times_reordered.}

\item{funs_reordered}{A list of functions for approximating the
cyclic trends of gene express levels for each gene, reordered by
cell_times_reordered.}

\item{mu_reordered}{Estimated cyclic trend of gene expression
values for each gene, reordered by cell_times_reordered.}

\item{prob_per_cell_by_celltimes}{Probabilities of each cell belong
to each bin.}
}
\description{
Apply the estimates of cycle_npreg_insample to another
gene expression dataset to infer an angle or cell cycle phase for
each cell.
}
