% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform_quantile.R
\name{data_transform_quantile}
\alias{data_transform_quantile}
\title{Quantile-normalize log counts of gene expression values}
\usage{
data_transform_quantile(Y, ncores = 2)
}
\arguments{
\item{Y}{A gene by sample matrix. Contains library size-normalized molecule counts.}

\item{ncores}{We use mclapply function for parallel computing.}
}
\value{
A gene by sample expression matrix.
}
\description{
For each gene, transform the log counts to a normal distribution.
  This way the zero-count cells are assigned the lowest qunatiles.
}
\author{
Joyce Hsiao
}
