% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform_quantile.R
\name{data_transform_quantile}
\alias{data_transform_quantile}
\title{Transform counts by first computing counts-per-million (CPM), then
  quantile-normalize CPM for each gene}
\usage{
data_transform_quantile(Y_sce, ncores = 2)
}
\arguments{
\item{Y_sce}{SingleCellExperiment Object.}

\item{ncores}{We use doParallel package for parallel computing.}
}
\value{
SingleCellExperiment Object with an added slot of cpm_quant,
    cpm slot is added if it doesn't exist.
}
\description{
For each gene, transform CPM to a normal distribution.
This way the zero-count cells are assigned the lowest qunatiles.
}
\examples{
# use our data
library(SingleCellExperiment)
data(sce_top101genes)

# normalize expression counts to counts per million
sce_normed <- data_transform_quantile(sce_top101genes, ncores=2)

plot(y=assay(sce_normed, "cpm_quant")[1,],
     x=assay(sce_normed, "umi_counts")[1,],
    xlab = "Before quantile-normalization",
    ylab = "After quantile-normalization")

}
\author{
Joyce Hsiao
}
