% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform_quantile.R
\name{data_transform_quantile}
\alias{data_transform_quantile}
\title{Quantile-normalize log counts of gene expression values}
\usage{
data_transform_quantile(Y, ncores = 2)
}
\arguments{
\item{Y}{A gene by sample matrix. Contains library size-normalized
molecule counts.}

\item{ncores}{We use doParallel package for parallel computing.}
}
\value{
A gene by sample expression matrix.
}
\description{
For each gene, transform the log counts to a normal distribution.
This way the zero-count cells are assigned the lowest qunatiles.
}
\examples{
# use our data
library(SingleCellExperiment)
data(sce_sub)

# normalize expression counts to counts per million
counts_normed<-t((10^6)*t(assay(sce_sub)[1:5,])/colData(sce_sub)$molecules)
counts_quant <- data_transform_quantile(counts_normed, ncores=2)

plot(x=counts_normed[1,], y=counts_quant[1,],
    xlab = "Before quantile-normalization",
    ylab = "After quantile-normalization")

}
\author{
Joyce Hsiao
}
