% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eset_final_sub}
\alias{eset_final_sub}
\title{Filtered gene expression data, along with summarized FUCCI intensities
and sample information. Contain the top 100 cyclic genes from our data.}
\format{An ExpressionSet object with 888 samples and top 100 cyclicl genes.
    \describe{
       \item{\code{theta}}{Inferred angles}
       \item{\code{exprs(est_final)}}{Molecule counts}
     }}
\usage{
eset_final_sub
}
\description{
Filtered gene expression data, along with summarized FUCCI intensities
and sample information. Contain the top 100 cyclic genes from our data.
}
\keyword{datasets}
