% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclic_many.R
\name{fit_cyclical_many}
\alias{fit_cyclical_many}
\title{Estimate Cyclic Trend For Several Genes Using Trendfiltering}
\usage{
fit_cyclical_many(Y, theta, polyorder = 2, ncores = 2)
}
\arguments{
\item{Y}{A matrix (gene by sample) of gene expression values. The
expression values are assumed to have been normalized and
transformed to the standard normal distribution.}

\item{theta}{A vector of cell cycle phase (angles) for single-cell
samples.}

\item{polyorder}{Argument passed to \code{\link{fit_trendfilter}}
specifying the degree of the polynomials used in nonparametric
trend filtering.}

\item{ncores}{Argument passed to
\code{\link[parallel]{makeCluster}} specifying the number of
threads.}
}
\value{
A list containing the following elements:

\item{predict.yy}{A matrix of predicted expression values at observed
  cell cycle.}

\item{cellcycle_peco_ordered}{A vector of predicted cell
  cycle. Values range between 0 to 2pi}

\item{cellcycle_function}{List of predicted cell cycle functions.}

\item{pve}{Vector of proportion of variance explained in each gene by
  the predicted cell cycle.}
}
\description{
For each gene, apply trend filtering as implemented in
  the genlasso package to estimate cell cycle. For more details, see
  \code{link{fit_trendfiltering}}.
}
\examples{
library(SingleCellExperiment)
data(sce_top101genes)

# Select top 10 cyclic genes.
sce_top10 <- sce_top101genes[order(rowData(sce_top101genes)$pve_fucci,
                                  decreasing=TRUE)[1:10],]
coldata <- colData(sce_top10)

# Get cell cycle phase based on FUCCI scores.
theta <- coldata$theta
names(theta) <- rownames(coldata)

# Normalize expression counts.
sce_top10 <- data_transform_quantile(sce_top10, ncores=2)
exprs_quant <- assay(sce_top10, "cpm_quantNormed")

# Order FUCCI phase and expression.
theta_ordered <- theta[order(theta)]
yy_ordered <- exprs_quant[, names(theta_ordered)]

fit <- fit_cyclical_many(Y=yy_ordered, theta=theta_ordered)

}
\seealso{
\code{\link{fit_trendfilter}}

\code{\link{fit_trendfilter}} for fitting one gene
  using trend filtering.
}
\author{
Joyce Hsiao
}
