% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cyclic_one.R
\name{fit_loess}
\alias{fit_loess}
\title{Use loess to estimate cyclic trends of expression values}
\usage{
fit_loess(yy, time)
}
\arguments{
\item{yy}{A vector of gene expression values for one gene. The
expression values are assumed to have been normalized and
transformed to standard normal distribution.}

\item{time}{A vector of angles (cell cycle phase).}
}
\value{
A list with one element, \code{pred.yy}, giving the
estimated cyclic trend.
}
\description{
Use loess to estimate cyclic trends of expression values
}
\examples{
library(SingleCellExperiment)
data(sce_sub)
coldata <- colData(sce_sub)

# cell cycle phase based on FUCCI scores
theta <- coldata$theta
names(theta) <- rownames(coldata)

# normalize expression counts to counts per million
counts_normed <- t((10^6)*t(assay(sce_sub)[1:5,])/colData(sce_sub)$molecules)
counts_quant <- data_transform_quantile(counts_normed, ncores=2)

# order FUCCI phase and expression
theta_ordered <- theta[order(theta)]
yy_ordered <- counts_quant[1,match(names(theta_ordered), colnames(counts_quant))]

fit <- fit_loess(yy_ordered, time=theta_ordered)

plot(x=theta_ordered, y=yy_ordered, pch=16, cex=.7, axes=FALSE,
  ylab="quantile-normalized expression values", xlab="FUCCI phase",
  main = "loess fit")
points(x=theta_ordered, y=fit$pred.yy, col="blue", pch=16, cex=.7)
axis(2)
axis(1,at=c(0,pi/2, pi, 3*pi/2, 2*pi),
  labels=c(0,expression(pi/2), expression(pi), expression(3*pi/2),
  expression(2*pi)))
abline(h=0, lty=1, col="black", lwd=.7)

}
\author{
Joyce Hsiao
}
