% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_cyclic_list}
\alias{genes_cyclic_list}
\title{Genes ordered by their cyclic trend (stronog to weak)
and sample information}
\format{A data frame
    \describe{
       \item{\code{order}}{The ordering of the genes according to proportion of
            variacne explained (large to small)}
       \item{\code{ensg}}{ENSG gene ID}
       \item{\code{pve}}{Proportion of variance explained in the expression values by the
            estimated cyclic trend}
     }}
\usage{
genes_cyclic_list
}
\description{
Genes ordered by their cyclic trend (stronog to weak)
and sample information
}
\keyword{datasets}
