% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_cyclic_list}
\alias{genes_cyclic_list}
\title{List of top 100 cyclic genes}
\format{A data frame with the following columns, in which the rows
are ordered according to the proportion of variance explained
(large to small):

\describe{

  \item{ensg}{ENSG gene ID.}

  \item{pve}{Proportion of variance explained in the expression
  values by the estimated cyclic trend.}
}}
\usage{
genes_cyclic_list
}
\description{
Top 100 cyclic genes ordered by their cyclic trend (strong to weak)
  across 888 samples analyzed in the study.
}
\keyword{data}
