% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.npreg.R
\name{initialize_cell_times}
\alias{initialize_cell_times}
\title{Initialize cell time estimates for the nonparametric approach}
\usage{
initialize_cell_times(Y, method.initialize.theta = c("pca", "uniform"), ...)
}
\arguments{
\item{Y}{gene by sample matrix}

\item{nbins}{Y is grouped to nbins between 0 to 2pi.}
}
\description{
Estimate cell times using PC1 and PC2. If \code{nbins} is specified,
  then group the estimated cell times to \code{nbins} groups.
}
