% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_5genes_predict}
\alias{model_5genes_predict}
\title{A SingleCellExperiment object}
\format{
A list with the following elements

\describe{
  \item{cellcycle_peco}{Predict cell cycle,
   the values ranged between 0 to 2pi}
 }
}
\description{
Pre-computed peco results. Applied
\emph{cycle_npreg_outsample} and results stored in
\emph{model_5genes_train} to predict cell cycle phase for
single-cell samples of NA19098. The predicted cell cycle is stored
as variable \emph{cellcycle_peco}.
}
\examples{
data(model_5genes_predict)

}
\keyword{data}
