% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_dist.R
\name{rotation}
\alias{rotation}
\title{Rotate Circular Variable to Minimize Distance Between
  Variables.}
\usage{
rotation(ref_var, shift_var)
}
\arguments{
\item{ref_var}{A vector of reference angles.}

\item{shift_var}{A vector of angles to be compared to ref_var.}
}
\value{
The transformed value of \code{shift_var} after rotation and
  shifting.
}
\description{
Rotate circular variable \code{shift_var} to minimize
  the distance between \code{ref_var} and \code{shift_var}. Because
  the origin of the cell cycle phases is arbitrary, we transform the
  angles prior to computing the distance (rotation and
  shifting). After this, one can apply \code{circ_dist} to compute
  the distance between the output value and \code{ref_var}.
}
\examples{

# Create a vector of angles.
theta_ref <- seq(0,2*pi, length.out=100)

# Shift the origin of theta_ref to pi.
theta_compare <- shift_origin(theta_ref, origin = pi)

# Rotate theta_compare in a such a way that the distance between
# theta_ref and theta_compare is minimized.
theta_compare_rotated <-
  rotation(ref_var=theta_ref, shift_var=theta_compare)

par(mfrow = c(1,2))
plot(x=theta_ref, y = theta_compare)
plot(x=theta_ref, y = theta_compare_rotated)

}
\author{
Matthew Stephens
}
