% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_top101genes}
\alias{sce_top101genes}
\title{Molecule Counts of the 101 Significant Cyclical Genes in the
  888 samples Analyzed in the Study.}
\format{
A \code{SingleCellExperiment} object with 888 samples and
  the 101 significant cyclic genes. Also, {sce_top101genes$theta}
  contains the inferred angles (FUCCI phase) for each cell.
}
\description{
A \code{SingleCellExperiment} object containing
  processed molecule count data (obtained after gene and sample
  filtering). The \code{colData} slot contains sample phenotype
  information.
}
\examples{
library(SingleCellExperiment)
data(sce_top101genes)
dim(sce_top101genes@colData)

}
\keyword{data}
