% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_origin.R
\name{shift_origin}
\alias{shift_origin}
\title{Shift Origin of Angles}
\usage{
shift_origin(phase, origin)
}
\arguments{
\item{phase}{A vector of angles (in radians).}

\item{origin}{the new origin of the angles.}
}
\value{
A vector of angles shifted to the new origin.
}
\description{
Shift the origin of the angles for visualization.
}
\examples{
# Make a vector of angles.
theta <- seq(0,2*pi, length.out=100)

# Shift the origin of theta to pi.
theta_shifted <- shift_origin(theta, origin = pi)

plot(x=theta, y = theta_shifted)

}
\author{
Joyce Hsiao
}
